python/requirements.txt
psycopg2>=2.9.10,<3.0.0
python-dotenv>=0.21.0,<1.0.0